4.5 change data type of price to numeric in products table

+++ db/migrate/20171009144437_remove_price_from_products.rb
@@ -1,5 +1,5 @@
 class RemovePriceFromProducts < ActiveRecord::Migration[5.1]
   def change
-    remove_column :products, :price, :decimal
+    remove_column :products, :price, :integer
   end
 end
